DCC-213 Changes to allow system-wide access filtering by using a custom Shiro Realm
This commit sets the groundwork, but no actual filtering is in place
yet. Filtering will be done using Shiro subjects. This will be passed to
the system using the HTTP SecurityContext. Currently Jersey 2.0 has a
bug that prevents this from working, but according to the Jersey JIRA it
was patched two days ago and should be fixed in the release this coming
Monday.